% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sroot.R
\name{sroot}
\alias{sroot}
\title{Compute s_root score}
\usage{
sroot(y, params, type = "r-Pareto")
}
\arguments{
\item{y}{observations}

\item{params}{model parameters}

\item{type}{model type, can be one of "r-Pareto"}
}
\value{
s_root score
}
\description{
Compute s_root score
}
\examples{
sroot(c(1,2,3),list("kappa"=1, "tau"=4))
}
